refactor(db sub): add processor to assign action and add folder for differ server
