+++ db_subscriber/bookmark.py
@@ -1,7 +1,6 @@
 import os
 import datetime
-from gql import gql, Client
-from gql.transport.aiohttp import AIOHTTPTransport
+from gql import gql
 
 def check_pick_exists(memberId, storyId, gql_client):
 
@@ -109,9 +108,6 @@ def bookmark_handler(content, gql_client):
 
 
 if __name__ == '__main__':
-    gql_endpoint = os.environ['GQL_ENDPOINT']
-    gql_transport = AIOHTTPTransport(url=gql_endpoint)
-    gql_client = Client(transport=gql_transport,
-                        fetch_schema_from_transport=True)
-
+    
     # print(bookmark_handler(content, gql_client))
+    print("done")
