+++ db_subscriber/follow.py
@@ -1,6 +1,4 @@
-import os
-from gql import gql, Client
-from gql.transport.aiohttp import AIOHTTPTransport
+from gql import gql
 
 
 def follow_handler(content, gql_client):
@@ -50,8 +48,6 @@ def follow_handler(content, gql_client):
 
 
 if __name__ == '__main__':
-    gql_endpoint = os.environ['GQL_ENDPOINT']
-    gql_transport = AIOHTTPTransport(url=gql_endpoint)
-    gql_client = Client(transport=gql_transport, fetch_schema_from_transport=True)
-
+    
     # print(follow_handler(content, gql_client))
+    print("done")
