+++ logging_subscriber/Dockerfile
